remove code specific to paper plotting in reporting.py instead move to plots.py
